+++ R/baConTree.R
@@ -17,17 +17,18 @@
 #'
 #' @importFrom purrr walk map_dbl
 #' @importFrom progressr progressor
+#' @importFrom Brobdingnag as.brob cbrob
 #' @export
 baConTree <- R6Class(
   "baConTree",
   inherit = ContextTree,
   public = list(
 
-#' @param data Either a vector with discrete data or a list of vectors.
-#' @param maximalDepth Depth of the maximal tree considered.
-#' @param alpha Hyperparameter for the Dirichlet prior distribution of probabilities.
-#' @param priorWeights A function to be evaluated at each node that returns
-#' its weight in the prior distribution.
+    #' @param data Either a vector with discrete data or a list of vectors.
+    #' @param maximalDepth Depth of the maximal tree considered.
+    #' @param alpha Hyperparameter for the Dirichlet prior distribution of probabilities.
+    #' @param priorWeights A function to be evaluated at each node that returns
+    #' its weight in the prior distribution.
     initialize = function(data, maximalDepth = 5, alpha, priorWeights) {
       super$initialize(data, maximalDepth)
       self$activateRoot()
@@ -38,24 +39,25 @@ baConTree <- R6Class(
       for(node in self$nodes) {
         node$extra$alpha <- rep(alpha, private$m)
       }
-      private$computeIntegratedDirichlet()
       for(node in self$nodes) {
-        node$extra$priorWeight <- priorWeights(node)
-        node$extra$logPriorWeight <- log(priorWeights(node))
+        node$extra$priorWeight <- as.brob(priorWeights(node))
+        node$extra$logPriorWeight <- log(node$extra$priorWeight)
       }
+      private$computeLogQ()
       private$preComputeRatios()
+      private$buildRecursiveSigma()
     },
 
 
-#' @param steps Number of steps to run the Metropolis Hastings algorithm for.
-#' @details
-#' This method supports progress monitoring via the **progressr** package.
-#' Users can wrap the function call in `with_progress()` to display a progress
-#' bar while the function executes. If no progress handler is registered, the
-#' function will run without showing progress.
-#'
-#' To enable progress, register a handler and wrap the function call in
-#' `with_progress()`.
+    #' @param steps Number of steps to run the Metropolis Hastings algorithm for.
+    #' @details
+    #' This method supports progress monitoring via the **progressr** package.
+    #' Users can wrap the function call in `with_progress()` to display a progress
+    #' bar while the function executes. If no progress handler is registered, the
+    #' function will run without showing progress.
+    #'
+    #' To enable progress, register a handler and wrap the function call in
+    #' `with_progress()`.
     runMetropolisHastings = function(steps){
       pb <- progressor(steps/10)
       if(is.null(private$chain)){
@@ -102,42 +104,59 @@ baConTree <- R6Class(
       }
     },
 
-#' @description Chain generated via Metropolis Hastings algorithm.
-#' @returns Gets the sampled chain stored.
+    #' @description Chain generated via Metropolis Hastings algorithm.
+    #' @returns Gets the sampled chain stored.
     getChain = function(){
       private$chain
+    },
+
+    #' @description Marginal likelihood of the data under the Bayesian context tree model.
+    #' @param log Logical. If `TRUE` (default), returns the log marginal likelihood
+    #' as a plain R `numeric`. If `FALSE`, returns the marginal likelihood as a
+    #' `brob` object.
+    #' @details
+    #' This is computed from `sigmaPosterior/sigmaPrior` at the root node,
+    #' which equals the sum over all trees of `prior(T) * p(data | T)`,
+    #' normalised by the prior partition function.
+    #' @returns A `numeric` (when `log = TRUE`) or a `brob` (when `log = FALSE`).
+    getMarginalLikelihood = function(log = TRUE){
+      root <- self$nodes[["*"]]
+      if(log){
+        as.numeric(log(root$extra$sigmaPosterior) - log(root$extra$sigmaPrior))
+      } else {
+        root$extra$sigmaPosterior / root$extra$sigmaPrior
+      }
     }
   ),
   private = list(
     alpha = numeric(0),
     iterations = 0,
     chain = NULL,
-    computeIntegratedDirichlet = function(){
+    computeLogQ = function(){
       for(node in self$nodes) {
         result <- lgamma(sum(node$extra$alpha)) -
           sum(lgamma(node$extra$alpha)) +
           sum(lgamma(node$extra$alpha + node$counts)) -
           lgamma(sum(node$extra$alpha + node$counts))
-        node$extra$marginalNodeLL <- result
+        node$extra$nodeLogQ <- result
+        node$extra$logPosteriorWeight <- node$extra$logPriorWeight +
+          node$extra$nodeLogQ
+        node$extra$PosteriorWeight <- exp(as.brob(node$extra$logPosteriorWeight))
       }
     },
 
-    preComputeRatios = function(){
-      for(node in self$nodes) {
-        node$extra$posteriorWeight <- node$extra$logPriorWeight +
-          node$extra$marginalNodeLL
-      }
 
+    preComputeRatios = function(){
       for(node in self$nodes){
         if(!node$isLeaf()){
           childrenPosteriorSum <- sum(map_dbl(self$nodes[node$getChildrenPaths()],
-                                              ~.x$extra$posteriorWeight))
+                                              ~.x$extra$logPosteriorWeight))
           node$extra$childrenPosteriorWeight <- childrenPosteriorSum
         } else {
           node$extra$childrenPosteriorWeight <- NA
         }
         node$extra$growPosteriorRatio <- node$extra$childrenPosteriorWeight -
-          node$extra$posteriorWeight
+          node$extra$logPosteriorWeight
       }
 
       for(node in self$nodes){
@@ -148,6 +167,27 @@ baConTree <- R6Class(
           node$extra$prunePosteriorRatio <- NA
         }
       }
+    },
+
+    buildRecursiveSigma = function(){
+      L <- self$getMaximalDepth()
+      nodes <- self$nodes[sapply(self$nodes, function(node) node$getDepth()) == L]
+      for(node in nodes){
+        node$extra$sigmaPrior <- as.brob(node$extra$priorWeight)
+        node$extra$sigmaPosterior <- node$extra$PosteriorWeight
+      }
+      L <- L - 1
+      while(L >= 0){
+        nodes <- self$nodes[sapply(self$nodes, function(node) node$getDepth()) == L]
+        for(node in nodes){
+          node_children <- self$nodes[node$getChildrenPaths()]
+          children_priorSigmas <- cbrobl(sapply(node_children, function(node) node$extra$sigmaPrior))
+          children_posteriorSigmas <- cbrobl(sapply(node_children, function(node) node$extra$sigmaPosterior))
+          node$extra$sigmaPrior <- node$extra$priorWeight + prod(children_priorSigmas)
+          node$extra$sigmaPosterior <- node$extra$priorWeight + prod(children_posteriorSigmas)
+        }
+        L <- L - 1
+      }
     }
   )
 )
@@ -155,7 +195,7 @@ baConTree <- R6Class(
 #' @importFrom ggplot2 scale_fill_gradient2
 #' @importFrom igraph V edge_attr<- ends E
 #' @export
-plot.baConTree = function(x, ...){
+plot.baConTree <- function(x, ...){
   ig <- x$igraph(activeOnly = FALSE)
   ends_mat <- ends(ig, E(ig))
   values <- V(ig)[ends_mat[, 2]]$prunePosteriorRatio
@@ -164,4 +204,8 @@ plot.baConTree = function(x, ...){
     geom_edge_diagonal0() +
     geom_node_label(aes(label = nodeLabel, fill = prunePosteriorRatio)) +
     scale_fill_gradient2(low = "red", high = "blue", midpoint = 0)
+}
+
+cbrobl <- function(x){
+  do.call(cbrob, unname(x))
 }